Change get_grid_cell/toggle_cell when not wrapping
The internal delegate functions called by these functions when the grid
does not wrap are changed to no longer cast systematically to unsigned
their signed coordinates parameters, but compare them to 0 first. The
result will be identical for all sensible grid sizes (i.e. all sizes
for which no dimension overflows in the negative domain), but this
handling of integer sign is cleaner.