Replace optimized bit exprs with clearer ones
After analyzing the code with an online compiler, it appears that using
x >> 3 for x / 8 and x & 7 for x % 8 has no effect: the division and
modulo expressions are always optimized to the bit shift and AND
versions (nearly identical bytecode with -O0, identical with -O2).
Since the division/modulo versions are more "mathematical" and clearer
(as far as i'm concerned) they are used instead of the bit expressions.
This also provides the opportunity to remove some parens around these
expressions, since div/mov have high operator precedence.