Fix some documentation issues
Improper Doxygen commands are fixed (\a and \e are replaced by \p when
referring to parameters and by \c when referring to code constants).
Rulestrings are referred to as rulestrings everywhere, not "Golly
format rule".
Verbs starting the description of a fucntion do not take an s (they are
infinitive or imperative, but not indicative).