Define a variable & pattern rule for .ccmd files
The new variable CCMD holds paths to files ending with .ccmd in the
objects directory. These files are the output of clang -MJ for each
source file. The variable is added as prerequisite to the COMPDB rule.

The new pattern rule matches .ccmd files in the objects directory; its
prerequisite is the source file corresponding to the matched file.
The commans of the rule are basically a copy of the compile rule, with
the compiler changed to clang -MJ.