Define an alias rule for the compialtion database file
The rule compdb is added, it is an empty rule with the compilation db
as prerequisite. its purpose is to make it more accessible to build the
file, without having to type its full name.
This rule is also added to the phony targets list as it does not
designate an actual file.