Update icon data embedded in code
The icon_data array defined in gridwindow.c is changed, for both icon
sizes.

The color depth is also changed from 16 to 32 bits, so the member type
of the array is also changed to maintain the equivalence 1 member = 1
pixel. This affects the depth and color channel masks parameters to
SDL_CreateRGBSurfaceFrom and the include comment to stdint.h.