Extract code-embedded icon data to separate files
The icon data array values for 16- and 64-pixel icons are moved in
standalone files in the data directory (because that is what these
values are: data, not, per se, code). For it to be accessible to the
initial source file, the data directory is added to the include path
(passed as argument to the compiler option -I or /I).