Change project structure for compield files
Object files and executables with debug symbols or optimizations or for
tests are generated in specific subdirectories of out and obj, release,
debug and test.

This will allow the different build modes to coexist in parallel
without interfering.

The variable for the release objects/executable are renamed to
explicitly refer to their mode.

The required object files for tests are drawn from the debug subdir, as
well as the object files used for the compilation database.