Remove build mode target-specific variable appends
This feature is cool and elegant, but it turns out to not be very
flexible: it is not possible to have the target-specific value when
making a particular file directly and explicitly, e.g. a debug obj file
will not be compiled with debugging symbols.

The target-specific values are thus injected in the recipe of the
object file compilation target corresponding to each build target
(because these are the last targets in the chain of prerequisites).